WIP part 2; Have YML serialization working
