+++ Ignition.Infrastructure/Pipelines/MapItem.cs
@@ -1,0 +1,8 @@
+﻿namespace Ignition.Infrastructure.Pipelines
+{
+	public class MapItem
+	{
+		public string FieldName { get; set; }
+		public string MapTo { get; set; }
+	}
+}
