+++ Ignition.Infrastructure/Pipelines/TemplateItem.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace Ignition.Infrastructure.Pipelines
+{
+	public class TemplateItem
+	{
+		public string TemplateName { get; set; }
+		public List<MapItem> MapItems { get; set; }
+	}
+}
