+++ Ignition.Infrastructure/Pipelines/TemplateMap.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Ignition.Infrastructure.Pipelines
+{
+	public class TemplateMap
+	{
+		public List<TemplateItem> TemplateItem { get; set; }
+	}
+}
