+++ Ignition.Infrastructure/Pipelines/YmlSettingsReader.cs
@@ -1,48 +1,26 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Web;
 using YamlDotNet.Serialization;
 using YamlDotNet.Serialization.NamingConventions;
 
 namespace Ignition.Infrastructure.Pipelines
 {
-    public class YmlSettingsReader
-    {
-	    public YmlSettingsReader()
-	    {
-			var reader = new StringReader(Document);
-			var deserializer = new Deserializer(namingConvention: new CamelCaseNamingConvention());
-			var maps = deserializer.Deserialize<TemplateMap>(reader);
+	public class YmlSettingsReader
+	{
+		public static void Main(string[] args)
+		{
+			var item = new YmlSettingsReader();
 		}
-
-	    private const string Document = @"---
-            TemplateMap:
-				- TemplateItem: 
-					-TemplateName: CallToAction
-					MapItems:
-						- FieldName: Heading
-						MapTo: Title
-						- FieldName: Subtitle
-						MapTo: Subheading
-			";
-    }
-}
-public class TemplateMap
-{
-	private List<TemplateItem> TemplateItems { get; set; }
-}
-
-public class MapItem
-{
-	public string FieldName { get; set; }
-	public string MapTo { get; set; }
-}
-
-public class TemplateItem
-{
-	public string TemplateName { get; set; }
-	public List<MapItem> MapItems { get; set; }
+		public TemplateMap TemplateMap { get; set; }
+		public YmlSettingsReader()
+		{
+			var reader = new StringReader(File.ReadAllText(HttpContext.Current.Server.MapPath("~/App_Config/FieldSettings.yml")));
+			var deserializer = new Deserializer();
+			TemplateMap = deserializer.Deserialize<TemplateMap>(reader);
+		}
+	}
 }
