+++ Ignition.Root/Ignition.Root.csproj
@@ -157,7 +157,9 @@
     <Content Include="App_Config\Include\Ignition.ContentEditor.config">
       <SubType>Designer</SubType>
     </Content>
-    <None Include="App_Config\FieldSettings.yml" />
+    <Content Include="App_Config\FieldSettings.yml">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
     <None Include="Properties\PublishProfiles\Ignition.Root.pubxml" />
     <Content Include="Properties\SlowCheetah\SlowCheetah.Transforms.targets" />
     <None Include="Web.CI.config">
